feat: Add status of empty for none search result
