+++ @antv/gatsby-theme-antv/site/components/APIDoc.module.less
@@ -1,0 +1,88 @@
+.require {
+  font-style: italic;
+  font-weight: 500;
+  font-size: 12px;
+  line-height: 14px;
+  color: rgba(0, 0, 0, 0.85);
+  opacity: 0.35;
+}
+.default {
+  margin-left: 12px;
+  font-style: italic;
+  font-weight: 500;
+  font-size: 12px;
+  line-height: 14px;
+  color: rgba(0, 0, 0, 0.85);
+  opacity: 0.35;
+}
+
+.type {
+  margin-left: 12px;
+  font-family: PingFangSC;
+  font-size: 10px;
+  line-height: 12px;
+  color: #000000;
+  opacity: 0.45;
+  display: inline-block;
+  background-color: #ccc;
+  padding: 4px;
+  border-radius: 4px;
+}
+
+.rootCollapse {
+  background: rgba(0, 0, 0, 0.04) !important;
+  border: none !important;
+  .hidden {
+    display: none;
+  }
+  .rootItem {
+    display: block;
+    border-bottom: none !important;
+    border-top: 1px solid #d9d9d94f;
+    :global {
+      .ant-collapse-content {
+        border: none;
+        padding-bottom: 8px;
+        .ant-collapse-content-box {
+          padding: 0px 16px;
+          .ant-collapse {
+            background-color: #fff;
+            .ant-collapse-item {
+              border-bottom: 1px solid #d9d9d96e;
+            }
+            &:last-child {
+              .ant-collapse-item {
+                border-bottom: none;
+              }
+            }
+          }
+        }
+      }
+    }
+  }
+}
+
+.docContent {
+  overflow-y: auto;
+  height: calc(100vh - 112px);
+}
+
+.designContent {
+  padding: 8px 26px;
+}
+
+.editOnGtiHubButton {
+  margin-left: 0.5em;
+  font-size: 16px;
+  color: #999;
+  vertical-align: middle;
+  position: relative;
+  top: -2px;
+}
+:global {
+  mark {
+    background-color: #b48af9;
+    padding: 0 3px;
+    border-radius: 3px;
+  }
+}
