+++ @antv/gatsby-theme-antv/site/components/PlayGround.module.less
@@ -1,5 +1,5 @@
 @import '~antd/es/style/themes/default.less';
-@editor-height: calc(100vh - 64px);
+@editor-height: calc(100vh - 72px);
 
 .playgroundCard {
   background-color: #fff;
