+++ @antv/gatsby-theme-antv/site/components/Tabs.module.less
@@ -85,6 +85,19 @@ ul.tabs {
     display: none;
   }
 }
+.resultNum {
+  background-color: #873cf45e;
+  border-radius: 50px;
+  min-width: 20px;
+  height: 20px;
+  padding: 0 6px;
+  color: #fff;
+  font-weight: 400;
+  font-size: 12px;
+  line-height: 20px;
+  white-space: nowrap;
+  text-align: center;
+}
 
 .tabExtra {
   position: absolute;
