+++ @antv/gatsby-theme-antv/site/templates/example.tsx
@@ -1,17 +1,9 @@
-import React, { useState, useEffect } from 'react';
-import Mark from 'mark.js';
+import React, { useState } from 'react';
+
 import { graphql, Link } from 'gatsby';
-import {
-  Layout as AntLayout,
-  Menu,
-  Tooltip,
-  Anchor,
-  Affix,
-  Collapse,
-} from 'antd';
+import { Layout as AntLayout, Menu, Anchor, Affix } from 'antd';
 import {
   createFromIconfontCN,
-  EditOutlined,
   MenuFoldOutlined,
   MenuUnfoldOutlined,
 } from '@ant-design/icons';
@@ -23,21 +15,19 @@ import { useMedia } from 'react-use';
 import RehypeReact from 'rehype-react';
 import Article from '../components/Article';
 import SEO from '../components/Seo';
-import Tabs from '../components/Tabs';
+
 import PlayGround from '../components/PlayGround';
 import NavigatorBanner from '../components/NavigatorBanner';
+import APIDoc from '../components/APIDoc';
 import CustomTag from '../components/CustomTag';
 import { capitalize } from '../utils';
 import { usePrevAndNext } from '../hooks';
-import { getGithubSourceUrl } from './document';
-import CollapseIcon from '../components/CollapseIcon';
+
 import styles from './markdown.module.less';
 
 const MenuIcon = createFromIconfontCN({
   scriptUrl: '//at.alicdn.com/t/font_470089_9m0keqj54r.js', // generated by iconfont.cn
 });
-
-const { Panel } = Collapse;
 
 const renderMenuItems = (edges: any[]) =>
   edges
@@ -128,11 +118,7 @@ export default function Template({
 }) {
   const { allMarkdownRemark, site } = data; // data.markdownRemark holds our post data
   const { edges = [] } = allMarkdownRemark;
-  const [collapseData, updateCollapseData] = useState<string[]>([]);
-  const [searchResult, updateSearchResult] = useState<string[]>([]);
-  const [searchQuery, updateSearchQuery] = useState<string>('');
-  const [r0ActiveKeys, updateR0ActiveKeys] = useState<string[]>(['r0-0']);
-  const [r2ActiveKeys, updateR2ActiveKeys] = useState<string[]>(['r2-0']);
+
   const edgesInExamples = edges;
   const pathWithoutTrailingSlashes = location.pathname.replace(/\/$/, '');
   const { node: markdownRemark } =
@@ -190,15 +176,7 @@ export default function Template({
     slug.startsWith(key),
   );
   const [openKeys, setOpenKeys] = useState<string[]>(defaultOpenKeys);
-  let activeTab = 'API' as 'API' | 'design';
-  let exampleRootSlug = slug;
-  if (/\/examples\/.*\/API$/.test(pathWithoutTrailingSlashes)) {
-    activeTab = 'API';
-    exampleRootSlug = exampleRootSlug.replace(/\/API$/, '');
-  } else if (/\/examples\/.*\/design$/.test(pathWithoutTrailingSlashes)) {
-    activeTab = 'design';
-    exampleRootSlug = exampleRootSlug.replace(/\/design$/, '');
-  }
+
   const { exampleSections = {}, allDemos = [] } = pageContext;
   const [prev, next] = usePrevAndNext();
 
@@ -384,93 +362,6 @@ export default function Template({
     </div>
   );
 
-  const genExtra = (options: {
-    require?: string;
-    type?: string;
-    default?: string;
-  }) => (
-    <>
-      {options?.require && (
-        <span className={styles.require}>{options?.require}</span>
-      )}
-      {options?.type && <span className={styles.type}>{options?.type}</span>}
-      {options?.default && (
-        <span className={styles.default}>{options?.default}</span>
-      )}
-    </>
-  );
-
-  const r0HandleChange = (activeKey: any) => {
-    updateR0ActiveKeys(activeKey);
-  };
-
-  const r2HandleChange = (activeKey: any) => {
-    updateR2ActiveKeys(activeKey);
-  };
-
-  useEffect(() => {
-    if (!exampleSections?.API) return;
-
-    if (searchResult.length > 0) {
-      updateCollapseData(searchResult);
-    } else {
-      updateCollapseData(exampleSections?.API.structure[0].children);
-    }
-  }, [exampleSections, searchResult]);
-
-  useEffect(() => {
-    if (!collapseData) return;
-    if (!searchQuery) {
-      // updateCollapseData(exampleSections?.API.structure[0].children);
-      console.log('none');
-    } else {
-      const context = document.querySelector('#apiStructure');
-      const instance = new Mark(context);
-      const reg = new RegExp(searchQuery, 'gi');
-      instance.markRegExp(reg);
-    }
-  }, [collapseData, searchQuery]);
-
-  const renderCollapse = () => {
-    return (
-      <div id="apiStructure">
-        <Collapse
-          expandIcon={({ isActive }) => (
-            <CollapseIcon rotate={isActive ? 90 : 0} type="caret-right" />
-          )}
-          className={styles.rootCollapse}
-          activeKey={r0ActiveKeys}
-          onChange={r0HandleChange}
-        >
-          {collapseData.map((node: any, key: any) => (
-            <Panel
-              key={`r0-${key}`}
-              header={node.title}
-              className={styles.rootItem}
-            >
-              {node.children.map((child: any, i: number) => (
-                <Collapse
-                  bordered={false}
-                  activeKey={r2ActiveKeys}
-                  onChange={r2HandleChange}
-                  key={`r1-${i}`}
-                >
-                  <Panel
-                    header={child.title}
-                    key={`r2-${i}`}
-                    extra={genExtra(child.options)}
-                  >
-                    {renderAst(child.content)}
-                  </Panel>
-                </Collapse>
-              ))}
-            </Panel>
-          ))}
-        </Collapse>
-      </div>
-    );
-  };
-
   const dispatchResizeEvent = () => {
     const e = new Event('resize');
     window.dispatchEvent(e);
@@ -493,58 +384,16 @@ export default function Template({
             />
           </>
         )}
-
-        <div className={styles.docPane}>
-          <Tabs
-            slug={exampleRootSlug}
-            title={frontmatter.title}
-            active={activeTab}
-            updateR0ActiveKeys={updateR0ActiveKeys}
-            updateR2ActiveKeys={updateR2ActiveKeys}
-            updateSearchResult={updateSearchResult}
-            updateSearchQuery={updateSearchQuery}
-            showTabs={{
-              API: !!exampleSections.API,
-              design: !!exampleSections.design,
-            }}
-            content={collapseData}
-          />
-          <div className={styles.docContent}>
-            {exampleSections.API &&
-            activeTab === 'API' &&
-            collapseData.length > 0
-              ? renderCollapse()
-              : null}
-            {exampleSections.design && activeTab === 'design' ? (
-              <div className={styles.designContent}>
-                <h1 className={styles.demoTtile}>
-                  {frontmatter.title}
-                  <Tooltip title={t('在 GitHub 上编辑')}>
-                    <a
-                      href={getGithubSourceUrl({
-                        githubUrl,
-                        relativePath,
-                        prefix: 'examples',
-                      })}
-                      target="_blank"
-                      rel="noopener noreferrer"
-                      className={styles.editOnGtiHubButton}
-                    >
-                      <EditOutlined />
-                    </a>
-                  </Tooltip>
-                </h1>
-                <div>{renderAst(htmlAst)}</div>
-                <div
-                  /* eslint-disable-next-line react/no-danger */
-                  dangerouslySetInnerHTML={{
-                    __html: exampleSections.design.node.html,
-                  }}
-                />
-              </div>
-            ) : null}
-          </div>
-        </div>
+        <APIDoc
+          slug={slug}
+          pathWithoutTrailingSlashes={pathWithoutTrailingSlashes}
+          relativePath={relativePath}
+          githubUrl={githubUrl}
+          frontmatter={frontmatter}
+          exampleSections={exampleSections}
+          renderAst={renderAst}
+          htmlAst={htmlAst}
+        />
       </SplitPane>
     </div>
   );
